Implemented motor wrappers. Seperated plugin system from sensor_stream
